requirements.txt
arrow==0.12.1
binaryornot==0.4.4
certifi==2018.1.18
chardet==3.0.4
click==6.7
cookiecutter==1.6.0
defusedxml==0.5.0
Django==2.0.3
django-allauth==0.35.0
django-ckeditor==5.4.0
django-crispy-forms==1.7.2
django-filter==1.1.0
django-js-asset==1.1.0
djangorestframework==3.8.2
future==0.16.0
idna==2.6
Jinja2==2.10
jinja2-time==0.2.0
Markdown==2.6.11
MarkupSafe==1.0
oauthlib==2.0.7
Pillow==5.1.0
pkg-resources==0.0.0
poyo==0.4.1
psycopg2==2.7.4
psycopg2-binary==2.7.4
Pygments==2.2.0
python-dateutil==2.7.0
python3-openid==3.1.0
pytz==2018.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
sorl-thumbnail==12.4.1
urllib3==1.22
whichcraft==0.4.1
